Collapse aar behaviour into aar convention plugin
