--- build-plugins/src/main/kotlin/AarCompatibility.kt
@@ -1,9 +1,0 @@
-import org.gradle.api.attributes.AttributeCompatibilityRule
-import org.gradle.api.attributes.CompatibilityCheckDetails
-import org.gradle.api.attributes.LibraryElements
-
-class AarCompatibility : AttributeCompatibilityRule<LibraryElements> {
-    override fun execute(details: CompatibilityCheckDetails<LibraryElements>) {
-        if (details.producerValue?.name == "aar") details.compatible()
-    }
-}
