+++ compose-constructor/plugin-runtime/build.gradle.kts
@@ -1,4 +1,5 @@
 plugins {
+    id("com.android.library")
     id("io.ashdavies.library")
 }
