+++ events-app/build.gradle.kts
@@ -1,6 +1,7 @@
 import com.android.build.api.dsl.VariantDimension
 
 plugins {
+    id("com.android.library")
     id("io.ashdavies.library")
 }
