+++ local-storage/build.gradle.kts
@@ -1,4 +1,5 @@
 plugins {
+    id("com.android.library")
     id("io.ashdavies.library")
     id("io.ashdavies.sql")
 }
